Attempt to make multiclass work for regression
No test yet

The implementation of regression in databunch is not great. I had made a
field for labels that are not integer, FloatLabels in databunch, and
those get used, but now I have to always read both int and float64
labels. It's not really that much data, as the labels are expected to be
a fair bit less than the features, so I'm keeping it as it is. Still,
something to keep in mind in case it turns out to take significant
memory.